refactor: SubItemInfo DTO를 DTOs 공통 위치로 이동
