+++ Core/Network/Sources/Network/DTOs/Daily/DailyWeeklyScoreRequest.swift
@@ -55,11 +55,3 @@ public struct DailyWeeklyScoreResponse: Decodable, Sendable {
     }
 }
 
-public struct SubItemInfo: Decodable, Sendable {
-    public let subItem: String
-    public let score: Double
-
-    public func toEntity() -> SubItemInfoEntity {
-        SubItemInfoEntity(subItem: subItem, score: score)
-    }
-}
