+++ Core/Network/Sources/Network/DTOs/SubItemInfo.swift
@@ -1,0 +1,15 @@
+//
+//  SubItemInfo.swift
+//  QRIZ
+//
+
+import QRIZUtils
+
+public struct SubItemInfo: Decodable, Sendable {
+    public let subItem: String
+    public let score: Double
+
+    public func toEntity() -> SubItemInfoEntity {
+        SubItemInfoEntity(subItem: subItem, score: score)
+    }
+}
